Move traffic_monitor nagios to generic location
